+++ TDEngine2/include/graphics/CDebugUtility.h
@@ -124,7 +124,7 @@ namespace TDEngine2
 				\param[in] color A text's color
 			*/
 
-			TDE2_API void DrawText(const TVector2& screenPos, F32 scale, const CU8String& str, const TColor32F& color) override;
+			TDE2_API void DrawText(const TVector2& screenPos, F32 scale, const std::string& str, const TColor32F& color) override;
 
 			/*!
 				\brief The method draws an axis-aligned cross with specified parameters