+++ TDEngine2/source/ecs/CUIElementsProcessSystem.cpp
@@ -252,7 +252,7 @@ namespace TDEngine2
 		pUIElementMeshData->ResetMesh();
 		
 		/// \note Transfer vertices from pFont->GenerateMesh into UIMeshData component
-		auto&& textMeshVertsData = pFont->GenerateMesh({ pLayoutData->GetWorldRect(), 1.0f, pLabelData->GetOverflowPolicyType(), pLabelData->GetAlignType() }, CU8String(pLabelData->GetText()));
+		auto&& textMeshVertsData = pFont->GenerateMesh({ pLayoutData->GetWorldRect(), 1.0f, pLabelData->GetOverflowPolicyType(), pLabelData->GetAlignType() }, pLabelData->GetText());
 
 		for (const TVector4& currVertex : textMeshVertsData.mVerts)
 		{