+++ CHANGELOG.md
@@ -7,7 +7,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
-## [0.5.27] 2022-01-23
+## [0.5.27] 2022-01-25
+
+### Changed
+
+- **CU8String** class is now just a namespace for helper functions to operate with UTF-8 encoded strings. **U8C** alias was
+replaced with strongly typed version called **TUtf8CodePoint**.
 
 ## [0.5.26] 2022-01-22